+++ pkg/client/notificationMock.go
@@ -5,14 +5,24 @@ import (
 )
 
 var _ Notification = (*NotificationMock)(nil)
+var _ NotificationV2 = (*NotificationMock)(nil)
 
-// NotificationMock mimics notification service - implements Notification interface
-type NotificationMock struct{}
+// NotificationMock mimics notification service - returns information about notififed users
+type NotificationMock struct {
+	counter *notifiedUsersCounter
+}
 
 func NewNotificationMock() *NotificationMock {
-	return &NotificationMock{}
+	return &NotificationMock{
+		counter: newNotifiedUsersCounter(),
+	}
 }
 
 func (c *NotificationMock) SendTemplated(templateKey, language string, payload map[string]interface{}, subscribers ...uuid.UUID) error {
+	c.counter.Count(templateKey, language, subscribers...)
 	return nil
+}
+
+func (c *NotificationMock) GetNotifiedUsers() NotifiedUsers {
+	return c.counter.GetStatus()
 }