+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Interface client.NotificationV2 and its implementations to enable tests in survey-svc
+
 ### Changed
 
 ### Deprecated